Move I;16 drawing tests to their own file
